+++ src/full.rs
@@ -4,18 +4,20 @@ use http::HeaderMap;
 use pin_project_lite::pin_project;
 use std::borrow::Cow;
 use std::convert::{Infallible, TryFrom};
+use std::marker::PhantomData;
 use std::pin::Pin;
 use std::task::{Context, Poll};
 
 pin_project! {
     /// A body that consists of a single chunk.
     #[derive(Clone, Copy, Debug)]
-    pub struct Full<D> {
+    pub struct Full<D, E = Infallible> {
         data: Option<D>,
+        _error: PhantomData<fn() -> E>,
     }
 }
 
-impl<D> Full<D>
+impl<D, E> Full<D, E>
 where
     D: Buf,
 {
@@ -26,16 +28,19 @@ where
         } else {
             None
         };
-        Full { data }
+        Full {
+            data,
+            _error: PhantomData,
+        }
     }
 }
 
-impl<D> Body for Full<D>
+impl<D, E> Body for Full<D, E>
 where
     D: Buf,
 {
     type Data = D;
-    type Error = Infallible;
+    type Error = E;
 
     fn poll_data(
         mut self: Pin<&mut Self>,
@@ -63,17 +68,20 @@ where
     }
 }
 
-impl<D> Default for Full<D>
+impl<D, E> Default for Full<D, E>
 where
     D: Buf,
 {
     /// Create an empty `Full`.
     fn default() -> Self {
-        Full { data: None }
+        Full {
+            data: None,
+            _error: PhantomData,
+        }
     }
 }
 
-impl<D> From<Bytes> for Full<D>
+impl<D, E> From<Bytes> for Full<D, E>
 where
     D: Buf + From<Bytes>,
 {
@@ -82,7 +90,7 @@ where
     }
 }
 
-impl<D> From<Vec<u8>> for Full<D>
+impl<D, E> From<Vec<u8>> for Full<D, E>
 where
     D: Buf + From<Vec<u8>>,
 {
@@ -91,7 +99,7 @@ where
     }
 }
 
-impl<D> From<&'static [u8]> for Full<D>
+impl<D, E> From<&'static [u8]> for Full<D, E>
 where
     D: Buf + From<&'static [u8]>,
 {
@@ -100,7 +108,7 @@ where
     }
 }
 
-impl<D, B> From<Cow<'static, B>> for Full<D>
+impl<D, E, B> From<Cow<'static, B>> for Full<D, E>
 where
     D: Buf + From<&'static B> + From<B::Owned>,
     B: ToOwned + ?Sized,
@@ -113,7 +121,7 @@ where
     }
 }
 
-impl<D> From<String> for Full<D>
+impl<D, E> From<String> for Full<D, E>
 where
     D: Buf + From<String>,
 {
@@ -122,7 +130,7 @@ where
     }
 }
 
-impl<D> From<&'static str> for Full<D>
+impl<D, E> From<&'static str> for Full<D, E>
 where
     D: Buf + From<&'static str>,
 {
@@ -139,13 +147,13 @@ mod tests {
     async fn full_returns_some() {
         let mut full = Full::new(&b"hello"[..]);
         assert_eq!(full.size_hint().exact(), Some(b"hello".len() as u64));
-        assert_eq!(full.data().await, Some(Ok(&b"hello"[..])));
+        assert_eq!(full.data().await, Some(Ok::<_, Infallible>(&b"hello"[..])));
         assert!(full.data().await.is_none());
     }
 
     #[tokio::test]
     async fn empty_full_returns_none() {
         assert!(Full::<&[u8]>::default().data().await.is_none());
-        assert!(Full::new(&b""[..]).data().await.is_none());
+        assert!(Full::<_, Infallible>::new(&b""[..]).data().await.is_none());
     }
 }